+++ src/Util.php
@@ -14,15 +14,15 @@ namespace Pronamic\WordPress\Pay\Gateways\Adyen;
  * Util
  *
  * @author  Remco Tolsma
- * @version 1.0.0
+ * @version 1.0.5
  * @since   1.0.0
  */
 class Util {
 	/**
 	 * Filter null.
 	 *
-	 * @param array $array Array to filter null values from.
-	 * @return array
+	 * @param array<int|string, mixed> $array Array to filter null values from.
+	 * @return array<int|string, mixed>
 	 */
 	public static function filter_null( $array ) {
 		return array_filter( $array, array( __CLASS__, 'is_not_null' ) );